+++ src/main/java/mju/iphak/maru_egg/question/docs/AdminQuestionTypeStatusControllerDocs.java
@@ -1,7 +1,5 @@
 package mju.iphak.maru_egg.question.docs;
 
-import java.util.List;
-
 import org.springframework.web.bind.annotation.RequestBody;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -9,7 +7,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import mju.iphak.maru_egg.question.dto.request.UpdateQuestionTypeStatusRequest;
-import mju.iphak.maru_egg.question.dto.response.QuestionTypeStatusResponse;
 
 @Tag(name = "Admin QuestionTypeStatus API", description = "어드민 질문타입 상태 관련 API 입니다.")
 public interface AdminQuestionTypeStatusControllerDocs {
@@ -23,9 +20,4 @@ public interface AdminQuestionTypeStatusControllerDocs {
 		@ApiResponse(responseCode = "200", description = "질문타입 상태 변경 성공")
 	})
 	void updateQuestionTypeStatus(@Valid @RequestBody UpdateQuestionTypeStatusRequest request);
-
-	@Operation(summary = "전체 질문타입과 상태 조회", description = "전체 질문타입과 상태를 조회합니다.", responses = {
-		@ApiResponse(responseCode = "200", description = "전체 질문타입과 상태 조회 성공")
-	})
-	List<QuestionTypeStatusResponse> getQuestionTypeStatus();
 }
